# BE/docs/swagger.yaml
swagger: "2.0"
info:
  title: IT4409 Jira-Clone API
  description: |
    Backend API cho hệ thống quản lý dự án kiểu Jira.
    Clean Architecture (Go) + PostgreSQL (Supabase).
    
    **Hướng dẫn xác thực:**
    1. Gọi `/api/auth/register` hoặc `/api/auth/login` để lấy JWT token.
    2. Nhấn nút "Authorize" ở góc trên, nhập: `Bearer <token>`.
    3. Tất cả API có khóa 🔒 đều yêu cầu JWT.
  version: "2.0"
  contact:
    name: IT4409 Team
basePath: /
schemes:
  - http
securityDefinitions:
  BearerAuth:
    description: "Nhập JWT token. Ví dụ: Bearer eyJhbGciOiJI..."
    type: apiKey
    name: Authorization
    in: header

# ════════════════════════════════════════════════
#  DEFINITIONS (Cấu trúc dữ liệu)
# ════════════════════════════════════════════════
definitions:
  Envelope:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      data: {}
...
  UserDTO:
    type: object
    properties:
      id:
        type: string
      email:
        type: string
      name:
        type: string
      avatar_url:
        type: string
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
...
  SprintPatch:
    type: object
    properties:
      name:
        type: string
      goal:
        type: string
      startDate:
        type: string
        format: date-time
      endDate:
        type: string
        format: date-time
...
paths:
...
  /api/users/me:
    get:
      tags: [Users]
      summary: Xem hồ sơ cá nhân
      security: [{BearerAuth: []}]
      produces: [application/json]
      responses:
        200:
          description: Thông tin người dùng hiện tại
          schema:
            allOf:
              - $ref: "#/definitions/Envelope"
              - properties:
                  data:
                    $ref: "#/definitions/UserDTO"
        401:
          description: Chưa đăng nhập
    patch:
      tags: [Users]
      summary: Cập nhật hồ sơ cá nhân
      security: [{BearerAuth: []}]
      consumes: [application/json]
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              name:
                type: string
              avatar_url:
                type: string
      responses:
        200:
          description: Cập nhật thành công
  /api/users/{userID}:
    get:
      tags: [Users]
      summary: Xem thông tin người dùng khác
      security: [{BearerAuth: []}]
      parameters:
        - in: path
          name: userID
          type: string
          required: true
      responses:
        200:
          description: Thông tin người dùng
        404:
          description: Không tìm thấy
  /api/users:
    get:
      tags: [Users]
      summary: Tìm kiếm người dùng
      security: [{BearerAuth: []}]
      parameters:
        - in: query
          name: search
          type: string
          required: true
          description: Từ khóa tìm kiếm (tên hoặc email)
        - in: query
          name: page
          type: integer
          default: 1
        - in: query
          name: per_page
          type: integer
          default: 20
      responses:
        200:
          description: Danh sách người dùng phù hợp

  # ── PROJECTS (Người B) ─────────────────────────
...
  /api/projects/{projectID}/sprints:
    post:
      tags: [Sprints]
      summary: Tạo sprint mới
      security: [{BearerAuth: []}]
      consumes: [application/json]
      produces: [application/json]
      parameters:
        - {in: path, name: projectID, type: string, required: true}
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name]
            properties:
              name: {type: string}
              goal: {type: string}
              startDate: {type: string, format: date-time}
              endDate: {type: string, format: date-time}
      responses:
        201:
          description: Sprint đã tạo
          schema:
            $ref: "#/definitions/Envelope"
    get:
      tags: [Sprints]
      summary: Liệt kê sprints trong project
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: projectID, type: string, required: true}
      responses:
        200:
          description: Danh sách sprints
          schema:
            $ref: "#/definitions/Envelope"
  /api/projects/{projectID}/backlog:
    get:
      tags: [Sprints]
      summary: Lấy backlog (issues chưa gán sprint)
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: projectID, type: string, required: true}
      responses:
        200:
          description: Danh sách issues trong backlog
          schema:
            $ref: "#/definitions/Envelope"
  /api/sprints/{sprintID}:
    get:
      tags: [Sprints]
      summary: Chi tiết sprint
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: sprintID, type: string, required: true}
      responses:
        200:
          description: Thông tin sprint
          schema:
            $ref: "#/definitions/Envelope"
    patch:
      tags: [Sprints]
      summary: Cập nhật sprint (partial)
      security: [{BearerAuth: []}]
      consumes: [application/json]
      produces: [application/json]
      parameters:
        - {in: path, name: sprintID, type: string, required: true}
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/SprintPatch"
      responses:
        200:
          description: Sprint đã cập nhật
          schema:
            $ref: "#/definitions/Envelope"
  /api/sprints/{sprintID}/start:
    post:
      tags: [Sprints]
      summary: Bắt đầu sprint (planning → active)
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: sprintID, type: string, required: true}
      responses:
        200:
          description: Sprint đã bắt đầu
          schema:
            $ref: "#/definitions/Envelope"
        409: {description: Đã có sprint đang active}
  /api/sprints/{sprintID}/complete:
    post:
      tags: [Sprints]
      summary: Hoàn thành sprint (active → completed)
      security: [{BearerAuth: []}]
      produces: [application/json]
      parameters:
        - {in: path, name: sprintID, type: string, required: true}
      responses:
        200:
          description: Sprint đã hoàn thành, issues chưa done chuyển về backlog
          schema:
            $ref: "#/definitions/Envelope"

  # ── ACTIVITY LOG (Người A) ─────────────────────
...
  /api/search:
    get:
      tags: [Search]
      summary: Tìm kiếm toàn cục (issues + projects)
      security: [{BearerAuth: []}]
      parameters:
        - {in: query, name: q, type: string, required: true, description: "Từ khóa tìm kiếm"}
        - {in: query, name: type, type: string, description: "Loại: issues, projects (mặc định tất cả)"}
        - {in: query, name: project_id, type: string, description: "Giới hạn trong dự án"}
      responses:
        200: {description: Kết quả tìm kiếm}

  # ── WEBSOCKET ──────────────────────────────────
...